[SB-36] Webapp > Repeat button
