+++ src/components/AudioPlayer/PlayerButtons/index.js
@@ -1,9 +1,9 @@
-import Repeat from './Repeat';
+import RepeatButton from './Repeat';
 import RewindButton from './Rewind';
 import PlayContainer from './Play';
 import ForwardButton from './Forward';
 import Shuffle from './Shuffle';
 
 export {
-  Repeat, RewindButton, PlayContainer, ForwardButton, Shuffle,
+  RepeatButton, RewindButton, PlayContainer, ForwardButton, Shuffle,
 };
