+++ src/store/constants.js
@@ -9,7 +9,7 @@ export const APP_TOKEN_UPDATE = 'APP_TOKEN_UPDATE';
 export const FILES_SET_FILES_LIST = 'FILES_SET_FILES_LIST';
 export const FILES_SET_FOLDER_LIST = 'FILES_SET_FOLDER_LIST';
 
-export const PLAYER_SET_REPEAT = 'PLAYER_SET_TOGGLE_REPEAT';
+export const PLAYER_SET_REPEAT = 'PLAYER_SET_REPEAT';
 export const PLAYER_SET_CURRENT_SONG = 'PLAYER_SET_CURRENT_SONG';
 export const PLAYER_SET_VOLUME = 'PLAYER_SET_VOLUME';
 export const PLAYER_SET_SONG_LINK = 'PLAYER_SET_SONG_LINK';
